Move error response into app namespace
