+++ src/lib.rs
@@ -74,8 +74,9 @@ pub mod utils;
 pub mod fixtures;
 
 pub use app::{
-    AppEvent, AppStoreData, Context, audit_log, authorised_agents, candidate_lists, candidates,
-    common, list_submitters, persons, political_groups, submit, substitute_list_submitters,
+    AppEvent, AppStoreData, Context, ErrorResponse, audit_log, authorised_agents, candidate_lists,
+    candidates, common, list_submitters, persons, political_groups, render_error_pages, submit,
+    substitute_list_submitters,
 };
 pub use auth::{
     derive_id::IdDeriver,
@@ -88,7 +89,7 @@ pub use core::{
     Province, TlsConfig, TypstRenderer, WaterCouncil, constants, get_env, http_trace, logging,
     server, translate,
 };
-pub use error::{AppError, AppResponse, ErrorResponse, render_error_pages};
+pub use error::{AppError, AppResponse};
 pub use form::{Form, TokenValue};
 pub use state::{AppRequestState, AppState};
 pub use utils::{
